fix: search 페이지 클라이언트에서 로드하도록 변경
- 서버에서 렌더 시 useSearchParam을 사용할 수 없음
