+++ src/app/search/page.tsx
@@ -1,37 +1,10 @@
-"use client";
+import SearchPage from "@/components/search/SearchPage";
+import { Suspense } from "react";
 
-import { useSearchParams, useRouter } from "next/navigation";
-import { useEffect } from "react";
-import SearchBar from "@/components/search/SearchBar";
-import CharacterCardContainer from "@/components/card/CharacterCardContainer";
-import "@/styles/globals.scss";
-import styles from "./SearchPage.module.scss";
-
-const SearchPage = () => {
-  const searchParams = useSearchParams();
-  const router = useRouter();
-  const keyword = searchParams.get("keyword")?.trim() ?? "";
-
-  useEffect(() => {
-    if (!keyword) {
-      router.push("/");
-    }
-  }, [keyword, router]);
-
-  if (!keyword) {
-    return <div>메인화면으로 돌아갑니다.</div>;
-  }
-
+export default function Page() {
   return (
-    <div className={styles.container}>
-      <main>
-        <div className={styles.searchSection}>
-          <SearchBar />
-        </div>
-        <CharacterCardContainer keyword={keyword} />
-      </main>
-    </div>
+    <Suspense>
+      <SearchPage />
+    </Suspense>
   );
-};
-
-export default SearchPage;
+}
